+++ code/src/cljs/sixsq/nuvla/ui/deployments/spec.cljs
@@ -41,4 +41,4 @@
                ::bulk-jobs-monitored     (sorted-map)
                ::bulk-jobs               (bulk-progress-plugin/build-spec)
                ::pagination              (pagination-plugin/build-spec
-                                           :default-items-per-page 8)})
+                                           :default-items-per-page 25)})