+++ code/src/cljs/sixsq/nuvla/ui/plugins/pagination.cljs
@@ -42,15 +42,16 @@
   {:content (r/as-element [ui/Icon {:name icon-name}]) :icon true})
 
 (defn Pagination
-  [{:keys [db-path total-items change-event] :as _opts}]
+  [{:keys [db-path total-items change-event i-per-page-multipliers] :as _opts}]
   (dispatch [::helpers/set db-path ::change-event change-event])
+(js/console.error "i-per-page-opts" i-per-page-multipliers)
   (let [dipp          @(subscribe [::helpers/retrieve db-path
                                    ::default-items-per-page])
         per-page-opts (map (fn [i]
                              {:key   (* dipp i)
                               :value (* dipp i)
                               :text  (* dipp i)})
-                           (range 1 4))
+                           (or i-per-page-multipliers (range 1 4)))
         change-page   #(dispatch [::change-page db-path %])
         tr            @(subscribe [::i18n-subs/tr])
         active-page   @(subscribe [::helpers/retrieve db-path ::active-page])
@@ -92,8 +93,13 @@
        :onPageChange  (ui-callback/callback :activePage #(change-page %))}]]))
 
 (s/def ::total-items (s/nilable nat-int?))
+(s/def ::i-per-page-multipliers (s/nilable #(and
+                                 (vector? %)
+                                 (nat-int? (first %))
+                                 (apply < %))))
 
 (s/fdef Pagination
         :args (s/cat :opts (s/keys :req-un [::helpers/db-path
                                             ::helpers/change-event
-                                            ::total-items])))
+                                            ::total-items
+                                            ::i-per-page-multipliers])))