+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - NuvlaEdge details - Hides owner dropdown in share tab, bugfix #998
 - NuvlaEdges table view - Table uses full width #1007
 - NuvlaEdges and Deployments - Makes table view default #1012
+- NuvlaEdges and Deployments - Sets default items per page to 25 with increasing items-per-page multiplier #1004
 
 ## [2.32.9] - 2022-11-17